[refactor] #45 : 공유모달 공통으로 분리
