+++ src/shared/model/index.ts
@@ -1,3 +1,4 @@
 export * from "./mocks/mockListData";
 export * from "./mocks/mockSearchData";
 export * from "./mocks/mockMapData";
+export * from "./kakaoShare.type";
