+++ src/shared/model/kakaoShare.type.ts
@@ -1,0 +1,6 @@
+export interface shareContentProps {
+  title: string;
+  description?: string;
+  imageUrl: string;
+  links: { label: string; url: string }[];
+}
