+++ src/shared/ui/index.ts
@@ -9,3 +9,4 @@ export * from "./BottomSheet/BottomSheet";
 export * from "./Portal";
 export * from "./Modal";
 export * from "./SnapBottomSheet/SnapBottomSheet";
+export * from "./ShareModal";
