+++ src/shared/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./validateName";
 export * from "./cookie";
+export * from "./kakaoShare";
